+++ src/Pim/Bundle/CatalogBundle/Manager/ProductManager.php
@@ -434,16 +434,6 @@ class ProductManager extends FlexibleManager
      *
      * @return ObjectRepository
      */
-    public function getAttributeOptionValueRepository()
-    {
-        return $this->entityManager->getRepository($this->getAttributeOptionValueName());
-    }
-
-    /**
-     * Return related repository
-     *
-     * @return ObjectRepository
-     */
     public function getFlexibleValueRepository()
     {
         return $this->entityManager->getRepository($this->getFlexibleValueName());